:recycle: Move entities to its own namespace
