+++ BattleshipGame/Models/Entities/Vessel.cs
@@ -1,8 +1,9 @@
-﻿using BattleshipGame.Models.Enumerations;
+﻿using BattleshipGame.Infrastructure.Models;
+using BattleshipGame.Models.Enumerations;
 
 namespace BattleshipGame.Models.Entities;
 
-public abstract class Vessel
+public abstract class Vessel : Entity
 {
     public string Name { get; set; }
     public int Width { get; set; }
