+++ BattleshipGame/Models/Entities/Vessels/AircraftCarrier.cs
@@ -1,7 +1,7 @@
 using BattleshipGame.Models.Enumerations;
 using Humanizer;
 
-namespace BattleshipGame.Models.Entities;
+namespace BattleshipGame.Models.Entities.Vessels;
 
 public class AircraftCarrier : Vessel
 {
