+++ BattleshipGame/Models/Entities/Vessels/Battleship.cs
@@ -1,7 +1,7 @@
 using BattleshipGame.Models.Enumerations;
 using Humanizer;
 
-namespace BattleshipGame.Models.Entities;
+namespace BattleshipGame.Models.Entities.Vessels;
 
 public class Battleship : Vessel
 {
