+++ BattleshipGame/Models/Entities/Vessels/Destroyer.cs
@@ -1,7 +1,7 @@
 using BattleshipGame.Models.Enumerations;
 using Humanizer;
 
-namespace BattleshipGame.Models.Entities;
+namespace BattleshipGame.Models.Entities.Vessels;
 
 public class Destroyer : Vessel
 {
